# Improve get task template detail