This PR isn’t a traditional single-query optimization; it makes a deliberate trade-off: accept a tiny overhead for small workflows (milliseconds to a few hundred ms) to speed up large workflows.

Original raw SQL:
SELECT DISTINCT 
    task_template.id,
    task_template.name,
    task_template_version.id AS task_template_version_id
FROM task
JOIN node ON task.node_id = node.id
JOIN task_template_version ON node.task_template_version_id = task_template_version.id
JOIN task_template ON task_template_version.task_template_id = task_template.id
WHERE 
    task.workflow_id = :workflow_id
    AND task_template_version.task_template_id = :task_template_id
    AND task_template_version.task_template_id = task_template.id
    
This has already been optimized. However, this performs poorly in corner cases. Example (Jon’s workflow 491686, task template 9740):
Set 1 (W): node_ids in the workflow → ~20K
Set 2 (T): node_ids that belong to the task template → ~2M
Joining to compute everything at once takes >12s in this case.

Indeed, we only need any one node in the overlap W ∩ T. For a given workflow and task template, all overlapping nodes share the same task_template_version_id, so finding one overlap is sufficient.

Approach (split into simpler steps)
1. Get the task template name (by :task_template_id).
2. Get Set 2 (T): node_ids associated with the task template.
3. Get Set 1 (W): node_ids in the workflow.
4: In Python, find one element of W ∩ T, then query its task_template_version_id.

Profiling (Jon’s example) using interactive SQL (note: this counts in the I/O time of interactive session):
Step 1: 0.004s
Step 2 (Set 2): 1.142s
Step 3 (Set 1): 0.040s
Step 4 (lookup id → version id): 0.013s
Python intersection cost:
Binary search in a 2M-element sorted list ≈ 0.01 ms
Worst case across 20K candidates ≈ 0.2 s
**Estimated total: ≲ 2 s.**

Production spot check using Jon’s workflow/template (query time only):
Old approach: 12.9 s
New approach: 0.02 s